Cambios 05/12/2022

+++ src/main/java/com/dboraclelab/entidades/Cliente.java
@@ -1,25 +1,25 @@
 package com.dboraclelab.entidades;
 
 public class Cliente {
-	private int cliente_id;
+	private long cliente_id;
 	private String nombres;
 	private String apellidos;
 	private boolean activo;
 	
 	public Cliente() {}
 	
-	public Cliente(int cliente_id, String nombres, String apellidos, boolean activo) {
+	public Cliente(long cliente_id, String nombres, String apellidos, boolean activo) {
 		this.cliente_id = cliente_id;
 		this.nombres = nombres;
 		this.apellidos = apellidos;
 		this.activo = activo;
 	}
 	
-	public int getCliente_id() {
+	public long getCliente_id() {
 		return cliente_id;
 	}
 	
-	public void setCliente_id(int cliente_id) {
+	public void setCliente_id(long cliente_id) {
 		this.cliente_id = cliente_id;
 	}
